moved redux files into components folder and updated files paths accordingly
