+++ client/src/components/ProjectsRedux.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import ProjectList from '../containers/ProjectList';
-import ProjectSearch from '../../utils/ProjectSearch';
-import { fetchAllProjects } from '../../actions';
-import FilterProject from '../../utils/FilterProjects';
+import ProjectList from './containers/ProjectList';
+import ProjectSearch from '../utils/ProjectSearch';
+import { fetchAllProjects } from '../actions';
+import FilterProject from '../utils/FilterProjects';
 import { Container } from 'react-bootstrap'
 
 class Projects extends Component {
