+++ client/src/components/ViewProfileProjectRedux.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import '../Style.scss';
-import ViewAProfileProject from '../containers/ViewAProfileProject';
+import './Style.scss';
+import ViewAProfileProject from './containers/ViewAProfileProject';
