+++ client/src/components/ViewProjectRedux.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Container } from 'react-bootstrap'
 import { connect } from 'react-redux';
-import '../Style.scss';
-import ViewAProject from '../containers/ViewAProject';
+import './Style.scss';
+import ViewAProject from './containers/ViewAProject';
